# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PaperMan",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/documents": {
      "get": {
        "summary": "Documents",
        "operationId": "documents",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Status"
            }
          },
          {
            "name": "delivery",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "",
                "review",
                "delivered"
              ],
              "type": "string",
              "default": "",
              "title": "Delivery"
            }
          },
          {
            "name": "inbox",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Inbox"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "date_desc",
                "date_asc",
                "title",
                "title_desc",
                "owners_asc",
                "owners_desc",
                "tags_asc",
                "tags_desc",
                "verification_asc",
                "verification_desc",
                "processed_asc",
                "processed_desc"
              ],
              "type": "string",
              "default": "date_desc",
              "title": "Sort"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspace": {
      "get": {
        "summary": "Workspace Settings",
        "operationId": "workspace_settings",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Workspace Preferences",
        "operationId": "workspace_preferences",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspacePreferences"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard": {
      "get": {
        "summary": "Dashboard",
        "operationId": "dashboard",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "running",
                    "review",
                    "failed",
                    "complete"
                  ],
                  "type": "string"
                },
                {
                  "const": "unverified",
                  "type": "string"
                }
              ],
              "default": "complete",
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings": {
      "get": {
        "summary": "Model Settings",
        "operationId": "settings",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettings-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Save Settings",
        "operationId": "save_settings",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelSettings-Input"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettings-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/rebuild": {
      "post": {
        "summary": "Rebuild",
        "operationId": "rebuild_index",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ActionResult": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ActionResult"
      },
...
      "Dashboard": {
        "properties": {
          "routing_documents": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "type": "array",
            "title": "Routing Documents"
          },
          "routing_total": {
            "type": "integer",
            "title": "Routing Total"
          },
          "documents": {
            "type": "integer",
            "title": "Documents"
          },
          "pending": {
            "type": "integer",
            "title": "Pending"
          },
          "review": {
            "type": "integer",
            "title": "Review"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "enrichment_failed": {
            "type": "integer",
            "title": "Enrichment Failed"
          },
          "worker": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkerState"
              },
              {
                "type": "null"
              }
            ]
          },
          "worker_online": {
            "type": "boolean",
            "title": "Worker Online"
          },
          "ocr_available": {
            "type": "boolean",
            "title": "Ocr Available"
          },
          "model_configured": {
            "type": "boolean",
            "title": "Model Configured"
          },
          "inbox_path": {
            "type": "string",
            "title": "Inbox Path"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          },
          "pipeline_items": {
            "$ref": "#/components/schemas/PipelinePage"
          },
          "unverified_documents": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "type": "array",
            "title": "Unverified Documents"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "routing_documents",
          "routing_total",
          "documents",
          "pending",
          "review",
          "failed",
          "enrichment_failed",
          "worker",
          "worker_online",
          "ocr_available",
          "model_configured",
          "inbox_path",
          "counts",
          "pipeline_items",
          "unverified_documents"
        ],
        "title": "Dashboard"
      },
      "Document": {
        "properties": {
          "owner_ids": {
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$"
            },
            "type": "array",
            "minItems": 1,
            "title": "Owner Ids"
          },
          "inbox_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,80}$",
            "title": "Inbox Id",
            "default": "shared"
          },
          "access_user_ids": {
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$"
            },
            "type": "array",
            "title": "Access User Ids"
          },
          "delivery_status": {
            "type": "string",
            "enum": [
              "review",
              "delivered"
            ],
            "title": "Delivery Status",
            "default": "review"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verification"
              },
              {
                "type": "null"
              }
            ]
          },
          "manual_rotations": {
            "items": {
              "$ref": "#/components/schemas/PageRotation"
            },
            "type": "array",
            "title": "Manual Rotations"
          },
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,80}$",
            "title": "Id"
          },
          "scan_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,80}$",
            "title": "Scan Id"
          },
          "source_pages": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Pages"
          },
          "processing_run": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Processing Run",
            "default": 1
          },
          "page_rotations": {
            "items": {
              "$ref": "#/components/schemas/PageRotation"
            },
            "type": "array",
            "title": "Page Rotations"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "document_date": {
            "type": "string",
            "format": "date",
            "title": "Document Date"
          },
          "date_source": {
            "type": "string",
            "enum": [
              "document",
              "scan_fallback"
            ],
            "title": "Date Source"
          },
          "scanned_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scanned At"
          },
          "final_path": {
            "type": "string",
            "title": "Final Path"
          },
          "processed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          },
          "user_tags": {
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$"
            },
            "type": "array",
            "title": "User Tags"
          },
          "excluded_tags": {
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$"
            },
            "type": "array",
            "title": "Excluded Tags"
          },
          "generated_tags": {
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$"
            },
            "type": "array",
            "title": "Generated Tags"
          },
          "suggested_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggested Tags"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "summary_edited": {
            "type": "boolean",
            "title": "Summary Edited",
            "default": false
          },
          "text_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Override"
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 0
          },
          "pdf_revision": {
            "type": "integer",
            "title": "Pdf Revision",
            "default": 0
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "type": "array",
            "title": "History"
          },
          "enrichment_status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "complete",
              "failed"
            ],
            "title": "Enrichment Status",
            "default": "pending"
          },
          "enrichment_version": {
            "type": "string",
            "title": "Enrichment Version",
            "default": ""
          },
          "enrichment_error": {
            "type": "string",
            "title": "Enrichment Error",
            "default": ""
          },
          "processing": {
            "items": {
              "$ref": "#/components/schemas/UsageAllocation"
            },
            "type": "array",
            "title": "Processing"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "owner_ids",
          "inbox_id",
          "access_user_ids",
          "delivery_status",
          "verification",
          "manual_rotations",
          "id",
          "scan_id",
          "source_pages",
          "processing_run",
          "page_rotations",
          "title",
          "document_date",
          "date_source",
          "scanned_at",
          "final_path",
          "processed_at",
          "user_tags",
          "excluded_tags",
          "generated_tags",
          "suggested_tags",
          "summary",
          "summary_edited",
          "text_override",
          "revision",
          "pdf_revision",
          "history",
          "enrichment_status",
          "enrichment_version",
          "enrichment_error",
          "processing"
        ],
        "title": "Document"
      },
...
      "DocumentPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "pages"
        ],
        "title": "DocumentPage"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelSettings-Input": {
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "compatible",
              "ollama",
              "demo"
            ],
            "title": "Provider",
            "default": "compatible"
          },
          "reasoning_effort": {
            "type": "string",
            "enum": [
              "default",
              "none",
              "low",
              "medium",
              "high"
            ],
            "title": "Reasoning Effort",
            "default": "default"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url",
            "default": ""
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 1800.0,
            "minimum": 5.0,
            "title": "Timeout Seconds",
            "default": 180
          },
          "output_mode": {
            "type": "string",
            "enum": [
              "prompted",
              "native",
              "tool"
            ],
            "title": "Output Mode",
            "default": "prompted"
          },
          "pricing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelPricing-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_before_filing": {
            "type": "boolean",
            "title": "Review Before Filing",
            "default": false
          },
          "ocr_languages": {
            "type": "string",
            "title": "Ocr Languages",
            "default": "eng"
          },
          "time_format": {
            "type": "string",
            "enum": [
              "24h",
              "12h"
            ],
            "title": "Time Format",
            "default": "24h"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ModelSettings"
      },
      "ModelSettings-Output": {
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "compatible",
              "ollama",
              "demo"
            ],
            "title": "Provider",
            "default": "compatible"
          },
          "reasoning_effort": {
            "type": "string",
            "enum": [
              "default",
              "none",
              "low",
              "medium",
              "high"
            ],
            "title": "Reasoning Effort",
            "default": "default"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url",
            "default": ""
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 1800.0,
            "minimum": 5.0,
            "title": "Timeout Seconds",
            "default": 180
          },
          "output_mode": {
            "type": "string",
            "enum": [
              "prompted",
              "native",
              "tool"
            ],
            "title": "Output Mode",
            "default": "prompted"
          },
          "pricing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelPricing-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_before_filing": {
            "type": "boolean",
            "title": "Review Before Filing",
            "default": false
          },
          "ocr_languages": {
            "type": "string",
            "title": "Ocr Languages",
            "default": "eng"
          },
          "time_format": {
            "type": "string",
            "enum": [
              "24h",
              "12h"
            ],
            "title": "Time Format",
            "default": "24h"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "reasoning_effort",
          "base_url",
          "model",
          "timeout_seconds",
          "output_mode",
          "pricing",
          "review_before_filing",
          "ocr_languages",
          "time_format"
        ],
        "title": "ModelSettings"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkspacePreferences": {
        "properties": {
          "time_format": {
            "type": "string",
            "enum": [
              "24h",
              "12h"
            ],
            "title": "Time Format"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "time_format"
        ],
        "title": "WorkspacePreferences"
      },
      "WorkspaceSettings": {
        "properties": {
          "time_format": {
            "type": "string",
            "enum": [
              "24h",
              "12h"
            ],
            "title": "Time Format"
          },
          "demo": {
            "type": "boolean",
            "title": "Demo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "time_format",
          "demo"
        ],
        "title": "WorkspaceSettings"
      }
    }
  }
}